Move spp to a meson subproject
