+++ .gitignore
@@ -59,20 +59,7 @@ librz/config.h
 librz/include/rz_userconf.h
 librz/include/rz_version.h
 librz/include/rz_version.h.tmp
-shlr/capstone/
-shlr/java/out
-shlr/java/out.exe
-shlr/sdb/sdb
-shlr/sdb/sdb.exe
-shlr/sdb/src/sdb
-shlr/sdb/src/.sdb
-shlr/sdb/src/sdb.exe
-shlr/sdb/src/.sdb.exe
-shlr/sdb/src/sdb-version.h
-shlr/sdb/src/sdb_version.h
-shlr/sdb/src/libsdb.so*
-shlr/spp/config.h
-shlr/.cs_tmp.zip
+subprojects/spp/config.h
 sys/travis/*.txt
 sys/.mark_python-deps
 sys/.mark_swig
