+++ librz/asm/meson.build
@@ -230,7 +230,7 @@ rz_asm = library('rz_asm', rz_asm_sources,
     rz_parse_dep,
     rz_bin_dep,
     capstone_dep,
-    spp_dep,
+    dependency('rzspp'),
   ],
   install: true,
   implicit_include_directories: false,
